requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-rest-knox==4.2.0
django-simple-history==3.2.0
psycopg2-binary==2.9.5
drf-yasg==1.21.5
python-dotenv==1.0.0
